client now deserializes 'generation' property to Long, was Int (#324)

+++ src/test/kotlin/com/redhat/devtools/intellij/kubernetes/model/resource/kubernetes/hasmetadata/HasMetadataResourceFactoryTest.kt
@@ -71,7 +71,7 @@ class HasMetadataResourceFactoryTest {
 		// metadata
 		val creationTimestamp = "creation1"
 		// deserizalization creates an Int
-		val generation: Int = 42
+		val generation: Long = 42
 		val name = "name1"
 		val namespace = "namespace1"
 		val metadataResourceVersion = "metadataResourceVersion1"
@@ -152,7 +152,7 @@ class HasMetadataResourceFactoryTest {
 
 	private fun createMetadataMap(
 			creationTimestamp: String? = null,
-			generation: Int? = null,
+			generation: Long? = null,
 			name: String? = null,
 			namespace: String? = null,
 			resourceVersion: String? = null,
